model, datalayer, tests

+++ Messenger/Messenger.DataLayer.Sql/UsersRepository.cs
@@ -20,32 +20,49 @@ namespace Messenger.DataLayer.Sql
                 connection.Open();
                 using (var transaction = connection.BeginTransaction())
                 {
+                    file.Id = Guid.NewGuid();
+                    user.Id = Guid.NewGuid();
+                    user.Photo = file.Id;
                     using (var command = connection.CreateCommand())
                     {
                         command.Transaction = transaction;
-                        command.CommandText = "insert into User (First_Name, Last_Name, Login, Password, Is_Active, Photo, User_Group) " +
-                                              "values ('" + user.FirstName + "', '" + user.LastName + "', '" +
-                                              user.Login + "', '" + user.Password + "', '" +
-                                              user.IsActive + "', '" + user.Photo + "', '" + user.UserGroup + "')";
+                        command.CommandText =
+                            "insert into Users (Id, FirstName, LastName, Password, IsActive) values (@Id, @FirstName, @LastName, @Password, @IsActive)";
+                        command.Parameters.AddWithValue("@Id", user.Id);
+                        command.Parameters.AddWithValue("@FirstName", user.FirstName);
+                        command.Parameters.AddWithValue("@LastName", user.LastName);
+                        command.Parameters.AddWithValue("@Password", user.Password);
+                        command.Parameters.AddWithValue("@IsActive", user.IsActive);
                         command.ExecuteNonQuery();
                     }
+                    
                     using (var command = connection.CreateCommand())
                     {
                         command.Transaction = transaction;
-                        file.Id = Guid.NewGuid();
                         command.CommandText = "insert into Files (Id, Name, Size, Owner)" +
-                                              "values ('" + file.Id + "', '" + file.Name + "', '" +
-                                              file.Size + "', '" + file.Owner + "' )";
+                                              "values (@Id, @Name, @Size, @Owner)";
+                        command.Parameters.AddWithValue("@Id", file.Id);
+                        command.Parameters.AddWithValue("@Name", file.Name);
+                        command.Parameters.AddWithValue("@Size", file.Size);
+                        command.Parameters.AddWithValue("@Owner", user.Id);
                         command.ExecuteNonQuery();
                     }
-                        transaction.Commit();
+                    using (var command = connection.CreateCommand())
+                    {
+                        command.Transaction = transaction;
+                        command.CommandText = "update Users set Photo=@FileId where Id=@Id";
+                        command.Parameters.AddWithValue("@FileId", file.Id);
+                        command.Parameters.AddWithValue("@Id", user.Id);
+                        command.ExecuteNonQuery();
+                    }
+                    transaction.Commit();
                 }
                 
             }
             return user;
         }
 
-        public void Delete(string login)
+        public void Delete(Guid id)
         {
             using (var connection = new SqlConnection(_connectionString))
             {
@@ -55,14 +72,21 @@ namespace Messenger.DataLayer.Sql
                     using (var command = connection.CreateCommand())
                     {
                         command.Transaction = transaction;
-                        command.CommandText = "delete from User where Login = '"+ login +"'";
+                        command.CommandText = "update Users set Photo=null";
                         command.ExecuteNonQuery();
                     }
                     using (var command = connection.CreateCommand())
                     {
                         command.Transaction = transaction;
-                        command.CommandText = "delete from Files " +
-                                              "where Id = (select Photo from User where Login = '"+ login +"') ";
+                        command.CommandText = "delete from Files where Owner=@Id";
+                        command.Parameters.AddWithValue("@Id", id);
+                        command.ExecuteNonQuery();
+                    }
+                    using (var command = connection.CreateCommand())
+                    {
+                        command.Transaction = transaction;
+                        command.CommandText = "delete from Users where Id =@userId";
+                        command.Parameters.AddWithValue("@userId", id);
                         command.ExecuteNonQuery();
                     }
                     transaction.Commit();
@@ -71,27 +95,27 @@ namespace Messenger.DataLayer.Sql
             }
         }
 
-        public User Get(string login)
+        public User Get(Guid id)
         {
             using (var connection = new SqlConnection(_connectionString))
             {
                 connection.Open();
                 using (var command = connection.CreateCommand())
                 {
-                    command.CommandText = "select * from User where Login = '" + login + "'";
+                    command.CommandText = "select * from Users where Id =@Id";
+                    command.Parameters.AddWithValue("@Id", id);
                     using (var reader = command.ExecuteReader())
                     {
                         if (!reader.Read())
-                            throw new ArgumentException($"Пользователь с Login {login} не найден");
+                            throw new ArgumentException($"Пользователь с Id {id} не найден");
                         return new User
                         {
-                            FirstName = Convert.ToString(reader["First_Name"]),
-                            LastName = Convert.ToString(reader["Last_Name"]),
-                            Login = Convert.ToString(reader["Login"]),
-                            IsActive = Convert.ToBoolean(reader["Is_Active"]),
-                            Photo = Convert.ToInt16(reader["Photo"]),
-                            Password = Convert.ToString(reader["Password"]),
-                            UserGroup = Convert.ToString(reader["User_Group"])
+                            FirstName = reader.GetString(reader.GetOrdinal("FirstName"))  ,
+                            LastName = reader.GetString(reader.GetOrdinal("LastName")),
+                            Id = reader.GetGuid(reader.GetOrdinal("Id")),
+                            IsActive = reader.GetInt32(reader.GetOrdinal("IsActive")),
+                            Photo = reader.GetGuid(reader.GetOrdinal("Photo")),
+                            Password = reader.GetString(reader.GetOrdinal("Password"))
                         };
                     }
                 }
@@ -106,11 +130,12 @@ namespace Messenger.DataLayer.Sql
                 using (var command = connection.CreateCommand())
                 {
                     command.CommandText =
-                        "update User(First_Name, Last_Name, Login, Password, Is_Active, User_Group)" +
-                        "values ('" + user.FirstName + "', '" + user.LastName + "', '" + user.Login + "', '"
-                        + user.Password + "', '" + user.IsActive + "', '" + user.UserGroup +
-                        "')" +
-                        "where Login = '" + user.Login + "'";
+                        "update Users set FirstName=@FirstName, LastName=@LastName, Password=@Password, IsActive=@IsActive where Id=@Id";
+                    command.Parameters.AddWithValue("@FirstName", user.FirstName);
+                    command.Parameters.AddWithValue("@LastName", user.LastName);
+                    command.Parameters.AddWithValue("@Password", user.Password);
+                    command.Parameters.AddWithValue("@IsActive", user.IsActive);
+                    command.Parameters.AddWithValue("@Id", user.Id);
                     command.ExecuteNonQuery();
                 }
